requirements.txt
# Cơ bản và an toàn cho Kaggle
wheel
timm==0.6.7
pytorch_metric_learning==1.2.0
pyyaml>=6.0
imageio>=2.9.0
imageio-ffmpeg>=0.4.5
yacs>=0.1.8
opencv-python-headless
motmetrics>=1.2.5
lap
cython
cython_bbox
python-dotenv
confluent-kafka==2.12.1
pyspark==3.5.6
pyarrow>=12.0.0
